backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
numpy
scikit-learn
psutil
python-dotenv
watchdog
PyMuPDF
python-docx
python-pptx
groq
pydantic
faster-whisper
pyttsx3
python-multipart
firebase-admin
